# Run pip in parallel for packages from requirements.txt to improve performance