Continuation of #204 - it looks like the difference for `pip` source is even greater:

2.6.0:
 - 56.279191s
 - 59.861275s
 - 54.958267s

2.6.0 + parallel (threads):
 - 16.790002s
 - 18.095694s
 - 18.228944s

Not all sources can benefit from the parallel run, but in the case of `pip` (and `pipenv`) as the script is executed for every package the improvement is noticeable.